Safari-609.1.11.1/LayoutTests/css3/flexbox/stretch-after-sibling-size-change.html

<!DOCTYPE html>
<html>
<head>
<link href="resources/flexbox.css" rel="stylesheet">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script>
window.onload = function()
{
    document.getElementById("a").style.height = "20px"
    checkLayout(".flexbox");
};
</script>
</head>
<body>
<div id=log></div>
<div data-expected-width="100" data-expected-height="20" class="flexbox" style="width: 100px">
    <div id=a class="flex-one" data-expected-width="50" data-expected-height="20" style="background-color: blue; height: 30px;"></div>
    <div id=b class="flex-one" data-expected-width="50" data-expected-height="20" style="background-color: green"></div>
</div>
</body>
</html>
